fix all

+++ datasetGen.py
@@ -51,7 +51,7 @@ class datasetGen:
         path = os.path.join(self.path_audio, os.listdir(self.path_audio)[file_num])
         
         # Perform the Constant-Q Transform
-        data, sr = librosa.load(path, sr = self.sr_downs, mono = True)
+        data, sr = librosa.load(path, sr = self.sr_downs, mono = True, dtype='float64')
         data = librosa.util.normalize(data)
         data = librosa.cqt(data, sr = self.sr_downs, hop_length = self.hop_length, fmin = None, n_bins = self.n_bins, bins_per_octave = self.bins_per_octave)
         CQT = np.abs(data)
